refactor(ts): split story.ts into focused submodules
story.ts (907 lines) split into: storyReader (types+constants+parsing), storySearch (full-text index), storyMemoir (chardict discovery), storySummary (event/chapter summaries). story.ts is now a re-export shim preserving all public symbols.

No behaviour change. typecheck clean, 138 tests pass.
